# perf(chunk): remove clone