--- a/packages/utilities/src/lib/chunk.ts
+++ b/packages/utilities/src/lib/chunk.ts
@@ -8,6 +8,6 @@ export function chunk<T>(array: readonly T[], chunkSize: number): T[][] {
 	if (!Number.isInteger(chunkSize)) throw new TypeError('chunkSize must be an integer.');
 	if (chunkSize < 1) throw new RangeError('chunkSize must be 1 or greater.');
 	const chunks: T[][] = [];
-	for (let i = 0; i < chunkSize; i += chunkSize) chunks.push(array.slice(i, i + chunkSize));
+	for (let i = 0; i < array.length; i += chunkSize) chunks.push(array.slice(i, i + chunkSize));
 	return chunks;
 }