--- a/packages/utilities/src/lib/chunk.ts
+++ b/packages/utilities/src/lib/chunk.ts
@@ -7,8 +7,7 @@ export function chunk<T>(array: readonly T[], chunkSize: number): T[][] {
 	if (!Array.isArray(array)) throw new TypeError('entries must be an array.');
 	if (!Number.isInteger(chunkSize)) throw new TypeError('chunkSize must be an integer.');
 	if (chunkSize < 1) throw new RangeError('chunkSize must be 1 or greater.');
-	const clone: T[] = array.slice();
 	const chunks: T[][] = [];
-	while (clone.length) chunks.push(clone.splice(0, chunkSize));
+	for (let i = 0; i < chunkSize; i += chunkSize) chunks.push(array.slice(i, i + chunkSize));
 	return chunks;
 }