… packages/eslint-config/tests/__snapshots__/eslint.test.ts.snap
… packages/prettier-config/tests/__snapshots__/prettier.test.ts.snap