Tested on an array with 21 elements:

| Version |   100,000 |      % | 1,000,000 |      % |
| :-----: | :-------: | :----: | :-------: | :----: |
| Old     | 257.704ms |   100% |    2.163s |   100% |
| New     | 162.308ms | 62.98% |    1.532s | 70.82% |

The updated function also uses half as much memory, as it doesn't clone the array. The performance improvement should be even greater on larger arrays, as removing (or inserting) elements at the start of an array is the slowest array operation.